Move components to be exported to `/react`
