+++ react/OrderOptions.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react'
 import { InjectedIntlProps, injectIntl } from 'react-intl'
 
-import ButtonLink from '../../ButtonLink'
+import ButtonLink from './ButtonLink'
 
 interface Props {
   allowCancellation: boolean
