+++ react/OrderTotal.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react'
 import { InjectedIntlProps, injectIntl } from 'react-intl'
 
-import Price from '../Payment/FormattedPrice'
+import Price from './FormattedPrice'
 
 interface Props {
   items: OrderItem[]
@@ -21,7 +21,7 @@ const ShippingTotals: FunctionComponent<Props & InjectedIntlProps> = ({
   const totalsLine = 'flex justify-between items-center c-muted-1'
 
   return (
-    <div className="flex-l flex-row justify-end w-100">
+    <div className="flex-l justify-end pl6-l w-100">
       <div className="mv8 w-100">
         <div className={totalsLine}>
           <p className="mv3">
