+++ react/PaymentMethod.tsx
@@ -1,10 +1,10 @@
 import React, { FunctionComponent, useState } from 'react'
 import { InjectedIntlProps, injectIntl } from 'react-intl'
 
-import ButtonLink from '../../ButtonLink'
-import InfoIcon from '../../Icons/Info'
-import AdditionalInfo from './AdditionalInfo'
+import ButtonLink from './ButtonLink'
+import AdditionalInfo from './components/Payment/AdditionalInfo'
 import Price from './FormattedPrice'
+import InfoIcon from './Icons/Info'
 
 interface Props {
   payment: Payment
